modified rpc_node.hpp
modified how the function name string is generated, now it generates off
a std::tuple<Ret, Args...> rather than just the std::tuple<Args...>,
this allows functions with the same set of arguments with a different
return type to exist at the same time.

NOTE: if for whatever reason two functions of the same arguments and
return type exists. the most recent registration will overwrite the
previous. so dont, i need to figure out a better way to handle that edge
case.