modified rpc_node.hpp
hash_code() is inconsistent depending on the compiler used. will use
just the typeid().name() until a better option is found.

also increased max func name length from 1024 to 65535 to accommedate
for the use of a name instead of a hash.

(maybe i can do an md5sum on the string and use that as a signature?)