netvar transparency
netvars can for the most part be assigned to transparently, or
constructed transparently. The only limitation is you have to modify the
whole variable at once, if the netvar is of a struct with multiple
members, it cannot be accessed individually. this may require C++
reflection which as of 2025 jan, is still not implemented